deploy update

+++ deploy.sh
@@ -1,7 +1,11 @@
 npm install
 npm uninstall governor-common
-npm install https://github.com/tinysquirrelton/governor-common.git
+npm install governor-common@git+https://github.com/Governor-DAO/governor-common-private.git
 RANDDIR="temp_build_$(openssl rand -hex 12)"
 BUILD_PATH=$RANDDIR npm run build
-rm -rf build
-mv $RANDDIR build
+if [ -f "$RANDDIR/index.html" ]; then
+    rm -rf build
+    mv $RANDDIR build
+else 
+    echo "Deploy failed!"
+fi
